Move jenkins jobs to their own directory, and normalize the names.
Summary:
We're going to be adding a lot of jenkins jobs if we move to
github-style deploys, and even if we don't we're starting to have a
lot of jobs.  We have given them all a build-* prefix so they're easy
to find in the tools directory, which sometimes make sense and
sometimes doesn't.  By having them all in their own directory, we
can name them all normally.

Test Plan:
None yet.  Once this is committed we'll have to change the jenkins
configs, and that will test all this. :-)

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D9656