Add a flag to automatically prune old versions when deploying a new one.
Summary:
One the big pain points for deploying, especially via jenkins, is you
start the process off and 2 minutes later it fails with 'too many
versions'.  Why can't deploy.py clean up the obsolete versions for
you, and then proceed?

The reason we didn't do this before is maybe we didn't trust the
auto-deleter.  But it's been working well for a long time now, so I
feel comfortable having it autodelete.

But I still only do it if you pass in an appropriate flag, which we
now do on jenkins.

Test Plan:
Manually ran
   deploy/deploy.py --clean-versions -fx -v znd-test
and saw it delete all the versions from 0606 and before, like it's
supposed to.

Reviewers: alpert, benkomalo

Reviewed By: alpert, benkomalo

Subscribers: alpert, chris

Differential Revision: http://phabricator.khanacademy.org/D9830