Add some awful hacks to deal with the fake tests that we run for parallel build.
Summary:
Our fake tests do 'make lint' via a python test
(testutil/manual_test.py:LintTest), because the build-parallelizer
wants only python tests.

So now when we are analyzing the output of the python tests, we notice
when a failing test is a fake test, and when so we simulate having run
the real test ('make lint') instead.

We abuse globals and write to files when it should be read-only.  I'm
so ashamed.

Test Plan:
Added a lint error to a file and ran
   tools/runtests.py --xml testutil.manual_test.LintTest.test_run_lint
   rm genfiles/lint_errors.txt    # just to be sure
   analyze_make_output.py --jenkins_build_url='foo' -n
and it said:
   INFO:root:alertlib: would send to hipchat room 1s and 0s: Failed 1 Lint check:<br>
   analyze_make_output.py:112:1: E302 expected 2 blank lines, found 1

Just like we wanted!  (Note it didn't report any python test failures.)

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D10140