Make build.lib sh-friendly.
We don't need any bash-isms in this script, so make it so sh scripts
can source it as well.  This requires only minor changes: 'source' ->
., and 'function' -> ().

(In theory, posix shells don't *need* to support the 'export FOO=bar'
idiom -- you are supposed to do 'FOO=bar; export FOO' instead -- but I
don't know of any shell that barfs on this, so I've kept those lines
as they are.)

Auditors: chris

Test Plan:
Fingers crossed.