Rewrite this script to take the branch-name as $1.
This makes it a lot easier to know what to check out -- it's only a
problem if two branches share the same sha1, but such a thing is
possible -- and also lets us simplify the logic somewhat.

I run in set +e mode now too, to better die on errors for things like
'git checkout'

Auditors: chris

Test Plan:
Fingers crossed.  Will run on jenkins once this is in.