First attempt at the back half of the deploy pipeline.
Summary:
This is the set-default stage, and the finish stage.  set-default was
already written but I revised it a lot to work better with
finish-deploy.  It is in charge of setting the default version on
appengine.  finish-deploy is responsible for rolling back to the old
default if there are problems, or else just releasing the deploy lock.

Both scripts spend a lot of time talking to hipchat.  I moved that
functionality into build.lib so other build scripts could easily talk
to hipchat as well.

Test Plan:
None yet.  Will deploy to jenkins and see how they work!  Hopefully
some way that doesn't involve actually setting the default.

Reviewers: chris

Differential Revision: http://phabricator.khanacademy.org/D10310