Time out after the old lock has been held for 60 minutes.
Before, our rule was, "wait for 60 minutes, then complain the old
deploy is taking too long."  But if the old deploy has already been
running for 3 hours, we should complain about too-long right away!  So
now we change the rule from 'wait for 60 minutes' to 'wait until the
old lock has been held for at least 60 minutes.'

I could have used the mtime of the props file rather than storing the
value in the props file, but I figured storing is more flexible (if we
ever want to move the props to a db or something).

Auditors: chris

Test Plan:
Ran
   deploy_pipeline.py acquire-lock --lockdir=$HOME/tmp/deploy.lockdir --deployer_username=csilvers --git_revision=deploy-test --auto_deploy=false --auto_rollback=false --jenkins_url=http://jenkins.khanacademy.org/ --hipchat_room=HipChat\ Tests --hipchat_sender=Mr\ Gorilla --deploy_email=prod-deploy@khanacademy.org --deploy_pw_file=/var/lib/jenkins/prod-deploy.pw

and then ran it again 2 minutes later and had it say:
   INFO:root:<b>csilvers</b>: You're next in line to deploy! (branch deploy-test.) Currently deploying (2 minutes in so far): csilvers