Replace 'git pull' with 'git checkout -B'.
I had added a 'git pull' to make sure the local branch is the same as
the remote branch: that is deploy-branch == origin/deploy-branch.  But
pull doesn't do that: it just makes sure deploy-branch >=
origin/deploy-branch.  If a previous deploy got into a weird state,
the local and remote branches could diverge.

So instead I run a command that makes the local and remote exactly the
same.  It turns out that 'git checkout -B' is a command that does
that.  So I run it now.

This simplifies how the code reads, as well.  It's a win-win.

Auditors: chris

Test Plan:
Will run from jenkins once this is checked in.