Use logging.xxx instead of print.
Jenkins doesn't do a good job of interleaving stdout and stderr,
leading to misleading output ordering, and delays while stdout is
buffered.  The solution is to write to stderr instead, and while I
could do 'print >>sys.stderr' everywhere, it's cleaner, and more
flexible, to use the logging module instead.

(I chose not to have the logging module print the date+time, since we
now have a jenkins plugin that does that for us, so it would just be
redundant info.)

Auditors: chris

Test Plan:
Ran
   deploy_pipeline.py finish-with-unlock --lockdir=$HOME/tmp/deploy.lockdir
and saw this:
   [ERROR] Running without having acquired the lock?