Recompute VERSION_NAME after merge-from-master.
VERSION_NAME depends on the sha1 that's associated with the branch
we're deploying, but merge-from-master can change that sha1 (due to
the merge it does).  So we need to recompute VERSION_NAME after the
merge.

To help clarify things, I now explicitly store in the props file the
sha1 associated with the deploy-branch at the time of deploy (so right
after the merge).  I also use that sha1 for future jenkins steps,
rather than the branch name, to make sure that all the deploy steps
are seeing *exactly* the same code.  Before, we could see variances if
someone pushed to the branch after the deploy-via-multijob step but
before the set-default step, say.

Auditors: chris

Test Plan:
Ran
   deploy_pipeline.py acquire-lock --lockdir=$HOME/tmp/deploy.lockdir --deployer_email=csilvers@khanacademy.org --git_revision=deploy-test --auto_deploy=false --auto_rollback=false --jenkins_url=http://jenkins.khanacademy.org/ --hipchat_room=HipChat\ Tests --hipchat_sender=Mr\ Gorilla --deploy_email=prod-deploy@khanacademy.org --deploy_pw_file=/var/lib/jenkins/prod-deploy.pw
and it said:
   [INFO] Wrote properties to /home/csilvers/tmp/deploy.lockdir: {'DEPLOYER_EMAIL': 'csilvers@khanacademy.org', 'AUTO_ROLLBACK': 'false', 'DEPLOY_PW_FILE': '/var/lib/jenkins/prod-deploy.pw', 'GIT_REVISION': 'deploy-test', 'JENKINS_URL': 'http://jenkins.khanacademy.org/', 'HIPCHAT_SENDER': 'Mr Gorilla', 'DEPLOYER_USERNAME': 'csilvers', 'AUTO_DEPLOY': 'false', 'DEPLOY_EMAIL': 'prod-deploy@khanacademy.org', 'ROLLBACK_TO': u'0620-1731-ae21870b0fe5', 'GIT_SHA1': 'deploy-test', 'LOCKDIR': '/home/csilvers/tmp/deploy.lockdir', 'DEPLOYER_HIPCHAT_NAME': u'@csilvers', 'VERSION_NAME': u'0624-1628-7f55a60bb79e', 'HIPCHAT_ROOM': 'HipChat Tests'}

Note VERSION_NAME is '0624-1628-7f55a60bb79e'.  Then I ran
   deploy_pipeline.py merge-from-master --lockdir=$HOME/tmp/deploy.lockdir
and it said:
   [INFO] Wrote properties to /home/csilvers/tmp/deploy.lockdir: {'LOCK_ACQUIRE_TIME': '1403653617', 'AUTO_DEPLOY': 'false', 'ROLLBACK_TO': '0620-1731-ae21870b0fe5', 'DEPLOY_EMAIL': 'prod-deploy@khanacademy.org', 'AUTO_ROLLBACK': 'false', 'DEPLOY_PW_FILE': '/var/lib/jenkins/prod-deploy.pw', 'DEPLOYER_EMAIL': 'csilvers@khanacademy.org', 'GIT_REVISION': 'deploy-test', 'GIT_SHA1': '131c70100dd01a820d166b5af50cb8c8fdefbb26', 'JENKINS_URL': 'http://jenkins.khanacademy.org/', 'HIPCHAT_ROOM': 'HipChat Tests', 'DEPLOYER_HIPCHAT_NAME': '@csilvers', 'LOCKDIR': '/home/csilvers/tmp/deploy.lockdir', 'DEPLOYER_USERNAME': 'csilvers', 'VERSION_NAME': u'0624-2357-131c70100dd0', 'HIPCHAT_SENDER': 'Mr Gorilla'}

and VERSION_NAME is now '0624-2357-131c70100dd0'.  A 'git log' shows
that this is indeed the commit-id of the merge.