Move the writing of the properties file to main().
merge_from_master has two places it does 'return', and we need to do
the _write_properties() for each of them.  Before, I was only doing it
from the second.  In general, we want the _write_properties() call to
happen whenever merge_from_master succeeds (or when acquire_lock
succeeds, for that matter), and the easiest way to ensure that is to
put the _write_properties() call outside merge_from_master().  Then I
don't have to worry about early returns.

Auditors: chris

Test Plan:
Fingers crossed.