Add the 'relock' command to re-acquire an accidentally deleted lock.
This will let us re-start a pipeline in the middle.  For instance, say
we run set-default on jenkins and it fails, but it was a spurious
failure and we want to run it again.  Right now we can't, because the
failure released the lock.  Now we can re-acquire the lock and run
set-default again.

Auditors: chris

Test Plan:
I ran
   deploy_pipeline.py acquire-lock --lockdir=$HOME/tmp/deploy.lockdir --deployer_email=csilvers@khanacademy.org --git_revision=deploy-test --auto_deploy=false --auto_rollback=false --jenkins_url=http://jenkins.khanacademy.org/ --hipchat_room=HipChat\ Tests --hipchat_sender=Mr\ Gorilla --deploy_email=prod-deploy@khanacademy.org --deploy_pw_file=/var/lib/jenkins/prod-deploy.pw
   deploy_pipeline.py finish-with-unlock --lockdir=$HOME/tmp/deploy.lockdir
   deploy_pipeline.py relock --lockdir=$HOME/tmp/deploy.lockdir.last

Then I ran 'tools/jenkins/deploy_pipeline.py acquire-lock [...]' and
it blocked since the lock is already acquired.

Then I ran
   deploy_pipeline.py finish-with-unlock --lockdir=$HOME/tmp/deploy.lockdir
   deploy_pipeline.py acquire-lock --lockdir=$HOME/tmp/deploy.lockdir --deployer_email=csilvers@khanacademy.org --git_revision=deploy-test --auto_deploy=false --auto_rollback=false --jenkins_url=http://jenkins.khanacademy.org/ --hipchat_room=HipChat\ Tests --hipchat_sender=Mr\ Gorilla --deploy_email=prod-deploy@khanacademy.org --deploy_pw_file=/var/lib/jenkins/prod-deploy.pw
   deploy_pipeline.py relock --lockdir=$HOME/tmp/deploy.lockdir.last
which (properly) failed with this error message:
   [ERROR] Cannot relock /home/csilvers/tmp/deploy.lockdir -- someone else has already acquired the lock since you released it.