Fix the set_default exceptions.
For some reason, I was using exceptions that suggest I was calling
set_default via subprocess, when in reality I am not doing that.  This
meant that when set_default failed I didn't catch it in the right way.

I revamped the exceptions and refactored the code to make it a bit
clearer what exception led to what follow-up code.  Other than the
name of the exception I look for, there should be no functional
change.

Auditors: chris

Test Plan:
Fingers crossed.