Move to a class-based, very java-y groovy helper class.
This way we can reuse code between jobs instead of having to cut and
paste the job everywhere.

This is probably the least groovy-ish code ever.  It's hard to find
good examples of the "right" way to do things!

Pair programmed with Chris.

Auditors: chris

Test Plan:
Will upload to jenkins and try it there.