Add CancelSiblings, similar but not identical to CancelDownstream.
Summary:
You can almost implement CancelSiblings by calling
   CancelDownstream(build.parent)
But that's not quite right because that would cause us to cancel
ourselves as well, and we don't want that (we want to keep our
existing build status, which will probably be 'failed').

If I were more comfortable with groovy I might try to extract out some
shared infrastructure, but I think this is fine for now.

Test Plan: Fingers crossed.  Will try this on jenkins once it's in!

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D10769