Always specify the remote we're pulling from or pushing to.
(In our case, it's always origin/master.)

When trying to pull from the new github-based intl/translations repo,
I got this:

   You asked to pull from the remote 'origin', but did not specify
   a branch. Because this is not the default configured remote
   for your current branch, you must specify a branch on the command line.

I have no idea why, but assume it has to do with how submodules are
set up.  There are various ways I could fix this by modifying the git
configuration, but the easiest way is to just change this script to be
more explicit.

Auditors: james, chris

Test Plan:
Fingers crossed; will run update-translations on jenkins once this is
in, and check the logs.