Replace hipchat-cli with alertlib, which does the same thing.
Summary:
The advantage is alertlib will use our existing secrets.py, so we
don't hae to copy the secret somewhere else for hipchat-cli to find
it.  (Also, we can now easily add logging to syslog/etc if we ever
wanted to.)

Once this is in, I'll also change the deploy-daily and deploy-znd
jenkins jobs to use alertlib rather than hipchat-cli.

Test Plan: Fingers crossed.  We'll find out if it works next deploy!

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D10931