Add an optional token to confirm that you own the lock.
Summary:
The idea is that when you create the lock, you pass in a random token,
and then you pass in the same token (if you want) to all subsequent
operations that need you to own the lock.  Then those operations not
only make sure that a lockfile is present, but that it has the same
value of the token that you do.  If not, it barfs.

We will use this in jenkins as an extra safety measure to make sure
that we don't accidentally do an operation on someone else's lock.

Test Plan:
Ran
   deploy_pipeline.py acquire-lock --lockdir=$HOME/tmp/deploy.lockdir --deployer_email=csilvers@ka.org --git_revision=deploy-token --auto_deploy=false --auto_rollback=false --jenkins_url=http://jenkins.khanacademy.org/ --hipchat_room=HipChat\ Tests --hipchat_sender=Mr\ Gorilla --deploy_email=prod-deploy@khanacademy.org --deploy_pw_file=/var/lib/jenkins/prod-deploy.pw
(no --token).  Then ran
   deploy_pipeline.py finish-with-unlock --lockdir=$HOME/tmp/deploy.lockdir --token=take1
and it unlocked.

Then I ran
   deploy_pipeline.py acquire-lock --lockdir=$HOME/tmp/deploy.lockdir --deployer_email=csilvers@ka.org --git_revision=deploy-token --auto_deploy=false --auto_rollback=false --jenkins_url=http://jenkins.khanacademy.org/ --hipchat_room=HipChat\ Tests --hipchat_sender=Mr\ Gorilla --deploy_email=prod-deploy@khanacademy.org --deploy_pw_file=/var/lib/jenkins/prod-deploy.pw --token=take1
   deploy_pipeline.py finish-with-unlock --lockdir=$HOME/tmp/deploy.lockdir
(no --token), and it unlocked.

Then I ran
   deploy_pipeline.py acquire-lock --lockdir=$HOME/tmp/deploy.lockdir --deployer_email=csilvers@ka.org --git_revision=deploy-token --auto_deploy=false --auto_rollback=false --jenkins_url=http://jenkins.khanacademy.org/ --hipchat_room=HipChat\ Tests --hipchat_sender=Mr\ Gorilla --deploy_email=prod-deploy@khanacademy.org --deploy_pw_file=/var/lib/jenkins/prod-deploy.pw --token=take1
   deploy_pipeline.py finish-with-unlock --lockdir=$HOME/tmp/deploy.lockdir --token=take2
and it failed to unlock with:
   [ERROR] You do not own the deploy lock (its token is take1, yours is take2); aborting

Then I ran
   deploy_pipeline.py finish-with-unlock --lockdir=$HOME/tmp/deploy.lockdir --token=take1
and it successfully unlocked.

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D10923