Wait for running jobs to actually finish before exiting.
Summary:
We'd like it to be an invariant that the cancel scripts don't return
until everything is actually canceled.  This way, if the jobs being
canceled do actions of their own upon cancelation, the callers of this
script can be confident those actions will have been taken by the time
the script returns.

Because of this rationale, we only need to do this waiting for
actively running jobs.  For jobs that we're deleting off the queue, we
don't need to wait, since such jobs never have a chance to do anything
(so nothing interesting can happen when they're canceled).

Sleep argument is ms, not seconds!

Test Plan:
Fingers crossed.  Will try on jenkins once this is in.

(same)

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D11117