Take out the 'wait for sibling to die before finishing' check.
Summary:
We had added code to CancelChildren which said, "Don't return until
all the children are actually dead."  For consistency, I added the
same code to CancelSiblings as well.  But for siblings it leads to
deadlock!  Sibling A tries to cancel sibling B, while simultaneously
sibling B is trying to cancel sibling A.  They both refuse to die,
though, until the other one is dead.

We don't actually *need* the 'wait before finish' behavior for
siblings.  For CancelChildren it makes sense, since we want to make
sure that the sub-jobs are actually finished before the parent job is
finished.  (This is also the only way to ensure that the parent job
can run steps in order.)

But for CancelSibling it doesn't really make sense, since you don't
"own" your sibling job in any way.  Here you're just hinting to the
sibling that it no longer has any purpose in life and may as well just
give up the candle.  But there's no real reason to wait until it does
so.

Test Plan: Fingers crossed.

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D11189