Fix some bugs, most notably one that created weird branch names.
I had coded
   git fetch origin +<branch>:<branch>
under the assumption that if <branch> is really a sha1, the fetch
would just fail (since there is no branch named <some_sha1>).  In git
1.8 that's true, but apparently in git 2.0 it succeeds, and creates
a local branch with the same name as the sha1.  Trouble!

So now I check if <branch> is actually a branch-name before doing the
fetch, using git ls-remote, which I just learned about while
researching this issue.

I made some other changes while in the area:

1. Instead of passing the GIT_SHA1 in to the next pipeline steps, pass in
the TOKEN.  The GIT_SHA1 was being totally ignored, while I've just
recently changed the pipeline jobs to use TOKEN if it exists.

This is important because it will avoid a potential future problem:
someone goes back into the dark past of hipchat history and clicks on
an old 'set default' link, and suddenly it sets default for the
currently pending deploy.  Now that the token will be wrong, that
re-click will properly be a noop.

2. When releasing someone else's lock (because they forgot to finish
the deploy themselves), just calling 'unlock' isn't good enough: we
need to call 'success' so they can merge into master, or 'rollback' if
there are serious problems (unlikely after such a long time, but...)
I now offer all of these as options.

Auditors: chris

Test Plan:
Fingers crossed.  I'm going to deploy this as soon as it's in, and see
how the process goes...