Bring deploy.sh into 2014 -- a pretty substantial rewrite.
Summary:
1. Add a new clean mode, some, which does what the old 'most' did.
2. Allow setting of more deploy.py flags, such as --modules and --skip-i18n
3. Let build.lib decrypt secrets instead of doing it ourselves.
4. Use alert (from build.lib) rather than calling alertlib ourselves.
5. Send hipchat messages as Mr Monkey, not Jenny Jenkins.
6. Don't keep popping in and out of the webapp dir.  It's annoying.

I changed the Makefile clean command to match what was here, which was
better (using git submodule foreach).

I also got rid of a commandline flag in deploy.py that wasn't being
used anywhere anymore.

I will have to change the deploy-znd jenkins job after this is in, to
pass through the new flags (such as NO_I18N), and to change the
default clean mode from 'most' to 'some'.

Test Plan:
Fingers way crossed.  I will do a znd deploy once this is checked in
(as a way of testing this code!)

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D11298