Emit a message as soon as we set-default, rather than waiting for monitoring.
Summary:
The message is like "we'll be monitoring for 10 minutes.  If you see
problems before then, you can cancel the deploy by clicking _here_."

This will speed things up for people, and also let them know that
something is happening even though it just seems like it's waiting.

Rename jenkins_job_url to jenkins_build_url.

This is for consistency with what jenkins calls this url (a job is
like 'make-check', a build is like 'make-check #245').  It also gives
consistency with analyze_make_output.py.

Test Plan:
Fingers crossed.  We'll see what happens on when I deploy it!

(same)

Reviewers: chris

Reviewed By: chris

Subscribers: tom

Differential Revision: http://phabricator.khanacademy.org/D11482