Store the last error found in the deploy, and use it in finish_with_failure.
Summary:
To make this easier, I moved the entire pipeline over to an
exception-based system: now instead of a pipeline function returning
True/False, it either returns or raises an exception.  The
exception-message is taken as the failure message, and future deploy
steps (presumably finish_with_failure or finish_with_rollback) will
look at that and pass it along to the user in a hipchat message.

This should make it a lot easier for people to see why a deploy
failed, without having to look at the jenkins logs.

Fix a typo, and make use of the auto-chaining in _update_properties.

Test Plan:
Ran
I changed merge_from_master to do this as the second line:

           raise RuntimeError('Merge conflict: must merge master into %s '
                           'manually.' % git_revision)

I then ran:
   deploy_pipeline.py acquire-lock --lockdir=$HOME/tmp/deploy.lockdir --deployer_email=csilvers@ka.org --git_revision=csilvers --auto_deploy=false --auto_rollback=false --jenkins_url=http://jenkins.khanacademy.org/ --hipchat_room=HipChat\ Tests --hipchat_sender=Mr\ Gorilla --deploy_email=prod-deploy@khanacademy.org --deploy_pw_file=/var/lib/jenkins/prod-deploy.pw
   deploy_pipeline.py merge-from-master --lockdir=$HOME/tmp/deploy.lockdir
   git co -    # because merge-from-master changes the branch to csilvers
   deploy_pipeline.py finish-with-failure --lockdir=$HOME/tmp/deploy.lockdir

and I saw that it wrote to hipchat:
   Deploy of 0728-1622-8c822eb7babb (branch csilvers) failed: Merge conflict: must merge master into csilvers manually.

I didn't test every function that I changed to return an exception
rather than True/False; we'll have to do that as part of the deploy
process.

(none new)

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D11671