Add support for a GENFILES_DIR envvar.
If set, we look for this directory, and if it exists use it as our
genfiles directory before deploying (we do this by mv-ing it).  This
is used when we want to split up the work of the deploy process,
having someone else do the building of genfiles (or at least most of
it) for us.

We will use this in conjunction with update-translations.  After it's
done updating translations, it will rebuild all the genfiles with the
new translations.  It will then provide the resulting genfiles-dir to
the deploy scripts so they don't have to spend time building the
new-translation genfiles themselves.

Pair programmed with james.

Auditors: james

Test Plan:
Fingers crossed.