Finish implementing AUTO_DEPLOY.
For some reason, I left it half-implemented: it did the set-default
automatically (as part of the deploy-via-mulitjob jenkins config), it
did not do the second manual step automatically, making it virtually
useless.  Now it does do the second step automatically.

On the other hand, if anything goes wrong with the set-default, I'm
now careful to always raise an exception in set_default() in
auto-deploy mode, to make sure we always roll back on error.
(AUTO_ROLLBACK is possibly redundant now, and we could get rid of it.)

I also avoid a bunch of hipchat messages around set-default when in
automatic mode, since they are telling people do something which is
not apprporiate during an auto-deploy.

Auditors: chris

Test Plan:
Fingers crossed.