Add a 'fast mv' function to build.lib.
This is used to rename a directory when the destination directory
might already exist, and we want to remove it before the rename.
The function does this removal very quickly, so the script doesn't
have to wait.  (It does this by procrastinating.)

Auditors: james

Test Plan:
Fingers crossed, but it matches code already working in
http://jenkins.khanacademy.org/job/update-translations/configure