Be more careful about the next steps for merge-from-master.
The next step for merge-from-master differs depending on some
user-specified options (e.g. AUTO_DEPLOY).  Before, I was setting
NEXT_STEPS to the union of all possibilities.  Now I look at the
user-specified options to determine the actual next step that should
be occurring for this deploy.

This avoids a problem where someone ran deploy-via-multijob and then
manually ran deploy-set-default before it was time to do so.  The
deploy-set-default job succeeded because 'set-default' was an
allowable NEXT_STEP after merge-from-master -- due to the fact it *is*
the next step when AUTO_DEPLOY is True -- even though in this
situation we should have waited until manual-test before we allowed
set-default.

Auditors: chris

Test Plan:
Fingers crossed