Fix a bug where we print the wrong username sometimes.
In general, we don't deal great with trying to run deploy-finish when
we don't have the lock (which can happen when deploy-via-multijob
fails when trying to acquire the lock).  I'll probably try to just
stop that from happening, but now at least it'll be a little less
confusing when it happens.

Auditors: chris

Test Plan:
Fingers crossed