Use 'git tag -f' in case we want to tag the same version bad twice.
Today there was a deploy where someone ended up deploying the exact
same code twice, and rolling it back each time.  The second time gave
an error because the 'git tag <sha1>-bad' failed since that tag
already existed.  Adding -f should let it succeed.  (It will replace
the tag contents with the exact same tag contents.)

This won't be a normal cause of events, but it can happen, and we
should be ready for it!

Auditors: chris

Test Plan:
Fingers crossed.