Send the relock instructions to hipchat, rather than just logging them.
Summary:
I didn't do this before because when we don't have the lockfile, we
don't have necessary information for logging, such as what hipchat
room to go to, what the jenkins url is, etc.  I get around that by
just guessing.  Even if it fails, or does the wrong thing, the error
message will still be in the logs like before (_alert() sends to logs
too), so this is strictly better than before.

Test Plan:
From the jenkins-tools directory, I ran (making sure the hipchat api
token wasn't available, so I didn't *actually* talk to hipchat!)
   % ./deploy_pipeline.py finish-with-failure --deployer_email=csilvers@khanacademy.org

And it said:
   [WARNING] Not sending this to hipchat (no token found): {'from': 'Mr Gorilla', 'message_format': 'text', 'color': 'red', 'room_id': '1s/0s: deploys', 'notify': 1, 'message': u'@csilvers (sadpanda) Trying to run without the lock. If you think you *should* have the lock, try to re-acquire it: http://jenkins.khanacademy.org/job/deploy-finish/parambuild?TOKEN=&STATUS=relock. Then run your command again.'}

I then visited
   http://jenkins.khanacademy.org/job/deploy-finish/parambuild?TOKEN=&STATUS=relock
and verified that this is the relock url.

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D13104