No longer build or check in the fake-language .po files (boxes.po, etc).
Summary:
Before, our approach was to build the fake-language .po files in
genfiles/, copy it to intl/translations, check it in, and then later
build the index.pickle and chunk.pickle files using the checked in
version.  Now we skip that middle step and just build the .pickle
files from the (e.g.) boxes.po that's in genfiles/.

(I think our old approach dated from the days when we used .po files
directly in production, and it made sense to have them all in a single
directory, so boxes.po was in the same dir as fr.po.  But those days
are long gone.)

This way we avoid having to check in boxes.po to github, which wasn't
working because the file got to be too big (github has a 100MB
per-file limit).

Test Plan: Fingers crossed.

Reviewers: james

Reviewed By: james

Differential Revision: http://phabricator.khanacademy.org/D13133