Check a smaller (sans comments) version of all.pot into intl/translations.
The 'big' all.pot is still stored in the crowdin repo, but this
smaller version will be useful for doing fake-translations, and also
maybe one day for doing incremental translations better (since it
keeps the string locations, so we can do a file-by-file update).

Auditors: james

Test Plan:
Manually ran
   % grep -v '^#\.' ~/khan/translations/all.pot > all.pot
where ~/khan/translations is the crowdin repo, and looked at the file
and it seemed reasonable.