Move from using a translations-data dir on kiln, to using on github.
Summary:
This used to not be practical, due to the limits that github imposes
on file size and repo size.  However, recent(ish) changes to the
translations process has changed the dynamics:

  * We download per-file translations files from crowdin, not all.zip
  * We don't bother with the .tmux stuff anymore

Now, everything fits nicely -- at least, as long as we keep all the
crowdin data compressed.  But there's no harm in doing that, it's just
historical.  (It may make git slower, since git doesn't deal well with
large binary files, but this repo is only used by these two jenkins
jobs, so it doesn't matter a huge amount if it's slow.)

If we run up against github limits again, it's easy to split this repo
into two, one for crowdin and one for captions.

Test Plan:
Fingers crossed.  Will run the two jenkins jobs manually after this is
checked in.

Reviewers: james

Reviewed By: james

Differential Revision: http://phabricator.khanacademy.org/D13263