Remove .rest.po and .datastore.po after downloading.
We had been removing them before downloading, but then
candidate_active_languages no longer has the .rest.po files it needs to
accurately determine the languages.  The main goal is to not resplit the
files and have es-ES.datastore.datastore.po or something.

Test Plan:
Made a test script with just that line.
touch intl/translations/es-ES.po
./test.sh
saw it print out all the .rest and .datasore, but not empty and not
es-ES.po

Auditors: csilvers