Fix update-translation breakage due to splitting non-existant en-PT.po.
Summary:
We were grabbing the list of languages to break in a different way than
when we did the downloading.  It saw a file that had en-PT, but it
wasn't en-PT.po.

I've moved the splitting now into the same loop as the downloading, so
we know the file must be there.  The difference is that now it will use
the old all.pot to determine what is in intl/datastore and what is in
rest.  This should be fine, because the download was based upon strings
from the previous upload's all.pot which is still around.  So in a way
it is better.

Test Plan: cross-fingers

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D14535