Print a 'stop' URL to allow users to cancel deploys early in the process.
Every so often someone will start a deploy and realize they want to
cancel it before the tests are even finished running.  Right now they
have to manually go to jenkins and hit the red 'x' on the deploy job.
Now we provide a URL to make it easier for them.  (Next step will be
to update hubot to support that url as well.)

I'm not sure how well this works: in theory, canceling the top-level
job should cancel the subjobs (build and deploy) as well, but I've not
actually tried it.  Jenkins can be a bit finicky about this.

A side effect is we now always post to hipchat when the deploy has
started, whereas before we didn't (we depended on the default jenkins
hipchat 'job has started' message).

Auditors: chris

Test plan:
Ran, in my webapp directory:
   env PYTHONPATH=.:tools ../jenkins-tools/deploy_pipeline.py acquire-lock --lockdir=$HOME/tmp/deploy.lockdir --deployer_email=csilvers@ka.org --git_revision=csilvers --auto_deploy=false --jenkins_url=http://jenkins.khanacademy.org/ --jenkins-build-url=http://jenkins.khanacademy.org/job/testjob/723/ --hipchat_room=HipChat\ Tests --hipchat_sender=Mr\ Gorilla --deploy_email=prod-deploy@khanacademy.org --deploy_pw_file=/var/lib/jenkins/prod-deploy.pw

and saw a reasonable message posted to hipchat:
   @csilvers Starting deploy of branch csilvers.  I'll post to HipChat when both a) tests are done and b) the deploy is finished.  If you wish to cancel before then:
   (failed) abort: http://jenkins.khanacademy.org/job/testjob/723/stop

I then ran it again without the --jenkins-build-url flag, and it said:
   @csilvers Starting deploy of branch csilvers.  I'll post to HipChat when both a) tests are done and b) the deploy is finished.

This seems right too.