Pass through the FORCE parameter from deploy-default.
Summary:
http://jenkins.khanacademy.org/job/deploy-default/configure passes
through FORCE=true if the user set DEPLOY=yes, and FORCE=false if the
user set DEPLOY=maybe.  We pass this through to default.py, which can
use it to decide when it is safe not to deploy (because this version
has already been deployed, say.)

Test Plan: Fingers crossed.

Reviewers: chris, benkomalo

Reviewed By: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D15036