Be better about downloading branches from the remote.
I was using
   git fetch origin +master:master
which is apparently not the way things are supposed to be done.
Instead you do
   git fetch origin +master:origin/master
followed by `git merge origin/master` or whatever to bring the
origin/master changes in with master.  (Though in our case we do `git
reset origin/master` since we don't have any local changes we want
merged in.)

The old way was giving us trouble when we were actually checked out
into the master branch:
   meta.1172 [master] webapp% git fetch origin +master:master
   fatal: Refusing to fetch into current branch refs/heads/master of non-bare repository

The problem being you can't update master's HEAD when you're already
in master.

I could fix this by doing something like
   git co HEAD^; git fetch origin +master:master; git co master
with the goal of getting out of the master branch before doing the
fetch, but that's just working around the bad behavior.  Better to do
what we're supposed to do and bring the refs over to origin/master
instead.

Auditors: chris

Test plan:
Fingers crossed.