Fix update-i18n-lite-videos by using a global lockfile for fetches.
Rather than trying to figure out exactly what repo we're fetching in
and acquiring the lock for just that repo, have a global lock that
applies to all fetches.  In practice this should be fine, and it was
getting hard to get the fine-grained locking working due to the
various ways we have to fetch in submodules.

update-i18n-lite-videos was doing this:
```
   safe_pull intl/translations
```
which cd's to the submodule and does a fetch, rather than the normal
way of doing a git submodule update, and that was throwing the locking
code for a loop.

Auditors: chris

Test plan:
Fingers crossed.  Will try to deploy update-i18n-lite-videos again
after this is in.