Change update-translations & sync-captions to save to dropbox instead of github
Summary:
Github complained that we were going over their 1GB limit claiming the
webapp-i18n-data account was using 30GB, so we need another place to
store the data.  Dropbox has two downsides, first we are no longer
storing the commit message along with the current sha.  Secondly we
can't so easily check out an old version.  Dropbox lets you download old
versions of a single file, but not an old version of an entire dir.  At
least things will be backed up ... just would be a pain to fully restore
things, but probably possible to write a script to get it via their api
or something.

On the plus side the script shouldn't die contacting git again.  Since
the syncing is done after the script finishes, we don't have to wait for
it to get anywhere, so it should finish quicker.

Test Plan:
Much of this was copied and tested on jenkins directly, as I rsynched
from the old webapp-i18n-data directory to new one.  The test for real
will happen after I cross-fingers and push.

Reviewers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D15657